Updating Script to use credentials file from google-auth action + secrets
